third_party/WebKit/LayoutTests/fast/events/touch/gesture/synthetic-pinch-zoom-gesture.html

<!DOCTYPE html>
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<meta name="viewport" content="width=device-width">
<style>
  body, html {
    width: 100%;
    height: 100%;
    margin: 0;
  }
  div {
    background-color: green;
    width: 100px;
    height: 100px;
    position: absolute;
    left: 50%;
    top: 50%;
    margin-left: -50px;
    margin-top: -50px;
  }
</style>

<script>
  const t = async_test(
      "This tests that gpuBenchmarking.pinchBy is relatively accurate.");

  function performZoom(scale, anchorX, anchorY, speed, callback) {
    chrome.gpuBenchmarking.pinchBy(scale, anchorX, anchorY, callback, speed);
  }

  const centerX = window.innerWidth / 2;
  const centerY = window.innerHeight / 2;

  // TODO(bokan): Add more variations (speed, anchor location, etc.) but it's
  // difficult right now as pinchBy is badly broken on desktops:
  // https://crbug.com/787615.
  const tests = [
    { starting_scale: 1, scale: 3, speed: 1000, x: centerX, y: centerY, msg: "Zooming in quickly is accurate" },
    { starting_scale: 4, scale: 0.5, speed: 1000, x: centerX, y: centerY, msg: "Zooming out quickly is accurate" },
  ];

  function runTest(testIx) {
    if (testIx > tests.length) {
      t.done();
      return;
    }

    const test = tests[testIx];
    window.internals.setPageScaleFactor(test.starting_scale);
    chrome.gpuBenchmarking.pinchBy(test.scale, test.x, test.y, () => {
      t.step(() => {
          assert_equals(window.visualViewport.scale, test.starting_scale * test.scale, test.msg);
          runTest(testIx + 1);
      });
    }, test.speed);
  }

  runTest(0);
</script>
